Update init order

+++ lua/core/plugins/yazi.lua
@@ -26,7 +26,7 @@ local M = {
   opts = {
     floating_window_scaling_factor = 0.90,
     open_for_directories = false,
-    yazi_floating_window_border = "double",
+    yazi_floating_window_border = "single",
     hooks = {
       yazi_opened_multiple_files = function(chosen_files, config, state)
         -- Iterate over the selected files
